python3/record-and-save-smb/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
requests==2.27.1
urllib3==1.26.8
